Split HDF5Writer into subfiles/classes
